requirements.txt

# # Core data processing
# pandas>=2.2.0
# numpy>=1.23.0,<2.0

# # Financial data
# yfinance==0.2.28
# pandas-datareader==0.10.0
# requests>=2.31.0

# # Scientific computing
# scipy>=1.11.1

# # Machine Learning
# scikit-learn>=1.3.0
# xgboost>=2.0.0
# torch>=2.0.0
# torchvision>=0.15.0

# # MLOps & Tracking
# mlflow>=2.9.0

# # Visualization
# matplotlib>=3.8.0
# seaborn>=0.13.0

# # Data validation
# great-expectations==0.17.23
# gcloud
# dvc==3.51.2
# dvc-gs==3.0.2
# xgboost
# shap

# # Utilities
# python-dotenv>=1.0.0
# gcloud
# joblib>=1.3.0



# optuna>=3.5.0
# plotly>=5.18.0
# optuna-integration[xgboost]>=3.5.0lightgbm==4.1.0




# Core data processing
pandas>=2.2.0
numpy>=1.23.0,<2.0

# Financial data
yfinance==0.2.28
pandas-datareader==0.10.0
requests>=2.31.0

# Scientific computing
scipy>=1.11.1

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
lightgbm==4.1.0

# MLOps & Tracking
mlflow>=2.9.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
shap

# Data validation
great-expectations==0.17.23
dvc==3.51.2
dvc-gs==3.0.2

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0


optuna>=3.5.0
plotly>=5.18.0
optuna-integration[xgboost]>=3.5.0
##lightgbm==4.1.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
lightgbm==4.1.0
# ============================================================================
# Additional packages for complete MLOps stack
# ============================================================================

# Weak Supervision (Model 3 - Anomaly Detection)
snorkel==0.9.9

# Imbalanced Learning
imbalanced-learn>=0.11.0

# Workflow Orchestration
apache-airflow>=2.7.0

# Model Serving & API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Configuration Management
PyYAML>=6.0.0

# Progress & Date Utilities
tqdm>=4.66.0
python-dateutil>=2.9.0

# Deep Learning Visualization
tensorboard>=2.20.0
tensorboard-data-server>=0.7.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
optuna>=3.5.0
plotly>=5.18.0
optuna-integration[xgboost]>=3.5.0
